--- coupon/src/test/kotlin/com/example/estdelivery/EstDeliveryCouponApplicationTests.kt
@@ -1,11 +1,0 @@
-package com.example.estdelivery
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class EstDeliveryCouponApplicationTests {
-    @Test
-    fun contextLoads() {
-    }
-}
