+++ coupon/src/test/kotlin/com/example/estdelivery/coupon/EstDeliveryCouponApplicationTests.kt
@@ -1,0 +1,11 @@
+package com.example.estdelivery.coupon
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class EstDeliveryCouponApplicationTests {
+    @Test
+    fun contextLoads() {
+    }
+}
